requirements.txt
langchain
langchain-classic
langchain-community
langchain-text-splitters
langchain-openai
langchain-chroma
pypdf
python-dotenv
